Automated Script for Encryption & Decryption
This is an Automated .sh Script that Encrypts & Decrypts a text file by choosing an option from the given list of choices. For Encryption, it uses GPG encryption.


Please note that: 

- Encrypt.sh   ---> it is the .sh file.
- encrypt.txt   ---> it is the sample text file that contains some data which is to be encrypted.
- encrypt.txt.gpg   ---> it is the GPG based encrypted file.


For Decryption, choose "Decryption" option and then it automatically shows output which is being stored in the text file.